-enh big refactor of irq handlers
