+++ CMakeLists.txt
@@ -16,17 +16,14 @@
 
 get_directory_property(PLATFORM_DEFINED COMPILE_DEFINITIONS)
 if (PLATFORM_DEFINED MATCHES "^__SAMD21")
-    add_library(Universal_hall "hal/platform/atmelsam/irq/default_irq_handlers.c"
+    add_library(Universal_hall 
+            "hal/platform/atmelsam/irq/irq_bindings.c"
             "hal/platform/atmelsam/gpio/gpio_samd.c"
             "hal/platform/atmelsam/i2c_host/i2c.c"
             "hal/platform/atmelsam/i2c_slave/i2c_slave.c"
             "hal/platform/atmelsam/spi_host/spi_host.c"
-            "hal/hal_spi_slave.h"
-            "hal/platform/atmelsam/spi_slave/spi_slave.c"
-            "hal/platform/atmelsam/spi_common/spi_platform_specific.h"
-            "hal/platform/atmelsam/clock_system/peripheral_clocking.h"
-            "hal/platform/atmelsam/irq/sercom_stuff.h"
-            "hal/hal_i2c_host.h")
+            "hal/platform/atmelsam/spi_slave/spi_slave.c" 
+            )
     target_include_directories(Universal_hall PUBLIC "hal/" "utils/" "hal/platform/atmelsam/")
     target_link_libraries(Universal_hall INTERFACE board_sdk)
 else ()
