+++ hal/platform/atmelsam/i2c_slave/i2c_slave.c
@@ -23,6 +23,7 @@
 */
 #include <stdbool.h>
 #include <hal_i2c_slave.h>
+#include "irq/irq_bindings.h"
 
 static Sercom *i2c_slave_peripheral_mapping_table[6] = {SERCOM0, SERCOM1, SERCOM2, SERCOM3, SERCOM4, SERCOM5};
 #define SERCOM_SLOW_CLOCK_SOURCE(x)               (x >> 8)
@@ -124,9 +125,9 @@ uhal_status_t i2c_slave_init(const i2c_periph_inst_t i2c_instance,
     NVIC_EnableIRQ(irq_type);
     const uint16_t irq_options = extra_configuration_options >> 8;
     if (irq_options) {
-        NVIC_SetPriority(irq_type, irq_options - 1);
+        enable_irq_handler(irq_type, irq_options - 1);
     } else {
-        NVIC_SetPriority(irq_type, 2);
+        enable_irq_handler(irq_type, 2);
     }
     return UHAL_STATUS_OK;
 }
